[#34] chore : add dto-under directory
- 하위 디렉토리 생성
