+++ BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/service/IssueService.java
@@ -6,11 +6,11 @@ import dev.codesquad.issuetracker.repository.IssueRepository;
 import dev.codesquad.issuetracker.repository.LabelRepository;
 import dev.codesquad.issuetracker.repository.MilestoneRepository;
 import dev.codesquad.issuetracker.repository.UserRepository;
-import dev.codesquad.issuetracker.web.dto.IssueResponse;
-import dev.codesquad.issuetracker.web.dto.MilestoneResponse;
+import dev.codesquad.issuetracker.web.dto.issue.IssueResponse;
+import dev.codesquad.issuetracker.web.dto.milestone.MilestoneResponse;
 import dev.codesquad.issuetracker.web.dto.ResultResponse;
 import dev.codesquad.issuetracker.web.dto.ResultDto;
-import dev.codesquad.issuetracker.web.dto.UserResponse;
+import dev.codesquad.issuetracker.web.dto.user.UserResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
