+++ BE/issue-tracker/src/main/java/dev/codesquad/issuetracker/web/dto/milestone/MilestoneDto.java
@@ -1,23 +1,23 @@
-package dev.codesquad.issuetracker.web.dto;
+package dev.codesquad.issuetracker.web.dto.milestone;
 
 import dev.codesquad.issuetracker.domain.milestone.Milestone;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
-public class MilestoneResponse {
+public class MilestoneDto {
 
     private Long id;
     private String title;
 
     @Builder
-    protected MilestoneResponse(Long id, String title) {
+    protected MilestoneDto(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public static MilestoneResponse of(Milestone milestone) {
-        return MilestoneResponse.builder()
+    public static MilestoneDto of(Milestone milestone) {
+        return MilestoneDto.builder()
             .id(milestone.getId())
             .title(milestone.getTitle())
             .build();
